+++ flow/project.py
@@ -555,10 +555,6 @@ class FlowProject(six.with_metaclass(_FlowProjectClass,
             flow_config.require_config_value
         template_environment.filters['get_account_name'] = tf.get_account_name
         template_environment.filters['print_warning'] = tf.print_warning
-        template_environment.filters['bold_font'] = tf.bold_font
-        template_environment.filters['draw_progressbar'] = tf.draw_progressbar
-        template_environment.filters['get_operation_status'] = tf.get_operation_status
-        template_environment.filters['job_filter'] = tf.job_filter
         if 'max' not in template_environment.filters:    # for jinja2 < 2.10
             template_environment.filters['max'] = max
         if 'min' not in template_environment.filters:    # for jinja2 < 2.10
@@ -1062,8 +1058,118 @@ class FlowProject(six.with_metaclass(_FlowProjectClass,
         if skip_active:
             raise NotImplementedError("The deprecated --skip-active option is no longer supported.")
 
+        # initialize jinja2 template evnronment and necessary filters
         env = self._environment
         template_environment = self._template_environment(env)
+
+        def draw_progressbar(value, total, width=40):
+            """Visualize progess with a progress bar.
+
+            :param value:
+                The current progress as a fraction of total.
+            :type value:
+                int
+            :param total:
+                The maximum value that 'value' may obtain.
+            :type total:
+                int
+            :param width:
+                The character width of the drawn progress bar.
+            :type width:
+                int
+            """
+
+            assert value >= 0 and total > 0
+            ratio = ' %0.2f%%' % (100 * value / total)
+            n = int(value / total * width)
+            return '|' + ''.join(['#'] * n) + ''.join(['-'] * (width - n)) + '|' + ratio
+
+        def job_filter(job_op, scheduler_status_code, all_ops):
+            """filter eligible jobs for status print.
+
+            :param job_ops:
+                Operations information for a job.
+            :type job_ops:
+                OrderedDict
+            :param scheduler_status_code:
+                Dictionary information for status code
+            :type scheduler_status_code:
+                Dictionary
+            :param all_ops:
+                Boolean value indicate if all operations should be displayed
+            :type all_ops:
+                Boolean
+            """
+
+            if scheduler_status_code[job_op['scheduler_status']] != 'U' or \
+               job_op['eligible'] or all_ops:
+                return True
+            else:
+                return False
+
+        def get_operation_status(operation_info, symbols):
+            """Determine the status of an operation.
+
+            :param operation_info:
+                Dicionary containing operation information
+            :type operation_info:
+                Dictionary
+            :param symbols:
+                Dicionary containing code for different job status
+            :type symbols:
+                Dictionary
+            """
+
+            if operation_info['scheduler_status'] >= JobStatus.active:
+                op_status = u'running'
+            elif operation_info['scheduler_status'] > JobStatus.inactive:
+                op_status = u'active'
+            elif operation_info['completed']:
+                op_status = u'completed'
+            elif operation_info['eligible']:
+                op_status = u'eligible'
+            else:
+                op_status = u'ineligible'
+
+            return symbols[op_status]
+
+        if pretty:
+            def highlight(s, eligible):
+                """Change font to bold within jinja2 template
+
+                :param s:
+                    The string to be printed
+                :type s:
+                    str
+                :param eligible:
+                    Boolean value for job eligibility
+                :type eligible:
+                    Boolean
+                """
+                if eligible:
+                    return '\033[1m' + s + '\033[0m'
+                else:
+                    return s
+        else:
+            def highlight(s, eligible):
+                """Change font to bold within jinja2 template
+
+                :param s:
+                    The string to be printed
+                :type s:
+                    str
+                :param eligible:
+                    Boolean value for job eligibility
+                :type eligible:
+                    boolean
+                """
+                return s
+
+        template_environment.filters['highlight'] = highlight
+        template_environment.filters['draw_progressbar'] = draw_progressbar
+        template_environment.filters['get_operation_status'] = get_operation_status
+        template_environment.filters['job_filter'] = job_filter
+
         template = template_environment.get_template(template)
         context = self._get_standard_template_context()
 
@@ -1185,7 +1291,6 @@ class FlowProject(six.with_metaclass(_FlowProjectClass,
         context['overview'] = overview
         context['detailed'] = detailed
         context['all_ops'] = all_ops
-        context['pretty'] = pretty
         context['parameters'] = parameters
         context['compact'] = compact
         context['unroll'] = unroll
@@ -1210,8 +1315,6 @@ class FlowProject(six.with_metaclass(_FlowProjectClass,
                 context['operation_status_symbols'] = OPERATION_STATUS_SYMBOLS
 
         print(template.render(**context), file=file)
-
-        return None
 
     def run_operations(self, operations=None, pretend=False, np=None, timeout=None, progress=False):
         """Execute the next operations as specified by the project's workflow.
