+++ flow/templates/print_status_compact.jinja
@@ -12,21 +12,18 @@
 {% endif %}
 {% if all_ops %}
 {% set key, value = job['operations'].items() | first() %}
-{{ field_job_id | format(job['job_id']) }}{{ field_operation | bold_font(value['eligible'], pretty) | format(key, '['+scheduler_status_code[value['scheduler_status']]+']', '+('+extra_num_operations | string()+')') }}{{para_output}}{{ '%s' | format(job['labels'] | join(', ')) }}
+{{ field_job_id | format(job['job_id']) }}{{ field_operation | highlight(value['eligible']) | format(key, '[' + scheduler_status_code[value['scheduler_status']] + ']', '+(' + extra_num_operations | string() + ')') }}{{para_output}}{{ '%s' | format(job['labels'] | join(', ')) }}
 {% else %}
 {% set ns.extra_num_operation = -1 %}
 {% set ns.if_first_eligible_operation = True %}
-{% for key, value in job['operations'].items() %}
-{% if (scheduler_status_code[value['scheduler_status']] != 'U' or value['eligible']) %}
+{% for key, value in job['operations'].items() if value | job_filter(scheduler_status_code, all_ops) %}
 {% set ns.extra_num_operation = ns.extra_num_operation + 1 %}
-{% if ns.if_first_eligible_operation %}
+{% if loop.first %}
 {% set ns.first_operation_key = key %}
 {% set ns.first_operation_value = value %}
-{% set ns.if_first_eligible_operation = False %}
-{% endif %}
 {% endif %}
 {% endfor %}
-{{ field_job_id | format(job['job_id']) }}{{ field_operation | bold_font(ns.first_operation_value['eligible'], pretty) | format(ns.first_operation_key, '['+scheduler_status_code[ns.first_operation_value['scheduler_status']]+']', '+('+ns.extra_num_operation | string()+')') }}{{para_output}}{{ '%s' | format(job['labels'] | join(', ')) }}
+{{ field_job_id | format(job['job_id']) }}{{ field_operation | highlight(ns.first_operation_value['eligible']) | format(ns.first_operation_key, '['+scheduler_status_code[ns.first_operation_value['scheduler_status']]+']', '+('+ns.extra_num_operation | string()+')') }}{{para_output}}{{ '%s' | format(job['labels'] | join(', ')) }}
 {% endif %}
 {% endfor %}
 {% endblock %}
