+++ flow/templates/print_status_expand.jinja
@@ -16,7 +16,7 @@
 {% for job in jobs %}
 {% for key, value in job['operations'].items() %}
 {% if (scheduler_status_code[value['scheduler_status']] != 'U' or value['eligible']) or all_ops%}
-{{ field_job_id | format(job['job_id']) }}{{ field_operation | bold_font(value['eligible'], pretty) | format(key) }}{{ '%-8s  ' | format(alias_bool[value['eligible']])}}{{ '%-14s  ' | format(scheduler_status_code[value['scheduler_status']])}}
+{{ field_job_id | format(job['job_id']) }}{{ field_operation | highlight(value['eligible']) | format(key) }}{{ '%-8s  ' | format(alias_bool[value['eligible']])}}{{ '%-14s  ' | format(scheduler_status_code[value['scheduler_status']])}}
 {% endif %}
 {% endfor %}
 {% endfor %}
