+++ flow/templates/print_status_stack.jinja
@@ -8,7 +8,7 @@
 {% endif %}
 {{ field_job_id | format(job['job_id']) }}{{para_output}}{{ '%s' | format(job['labels'] | join(', ')) }}
 {% for key, value in job['operations'].items() if value | job_filter(scheduler_status_code, all_ops) %}
-{{ field_operation | bold_font(value['eligible'], pretty) | format(value | get_operation_status(operation_status_symbols), key, '['+scheduler_status_code[value['scheduler_status']]+']') }}
+{{ field_operation | highlight(value['eligible']) | format(value | get_operation_status(operation_status_symbols), key, '['+scheduler_status_code[value['scheduler_status']]+']') }}
 {% endfor %}
 {% endfor %}
 {{ operation_status_legend }}
